requirements.txt
asgiref==3.3.1
cairocffi==1.2.0
CairoSVG==2.5.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cssselect2==0.4.1
defusedxml==0.6.0
Django==3.1.3
django-paypal==1.0.0
fpdf==1.7.2
idna==2.10
Pillow==8.0.1
pycparser==2.20
pytz==2020.4
qrcode==6.1
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
tinycss2==1.1.0
urllib3==1.26.2
webencodings==0.5.1
